ai_services/hoover4_ai_clients/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
anyio==4.10.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
certifi==2025.8.3 ; python_version >= "3.9" and python_full_version < "4.0.0"
charset-normalizer==3.4.3 ; python_version >= "3.9" and python_full_version < "4.0.0"
colorama==0.4.6 ; python_version >= "3.9" and python_full_version < "4.0.0" and sys_platform != "win32" and platform_system == "Windows"
exceptiongroup==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
grpcio==1.74.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
h11==0.16.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
httpcore==1.0.9 ; python_version >= "3.9" and python_full_version < "4.0.0"
httpx==0.28.1 ; python_version >= "3.9" and python_full_version < "4.0.0"
idna==3.10 ; python_version >= "3.9" and python_full_version < "4.0.0"
jsonpatch==1.33 ; python_version >= "3.9" and python_full_version < "4.0.0"
jsonpointer==3.0.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
langchain-core==0.3.76 ; python_version >= "3.9" and python_full_version < "4.0.0"
langsmith==0.4.27 ; python_version >= "3.9" and python_full_version < "4.0.0"
milvus-lite==2.5.1 ; python_version >= "3.9" and python_full_version < "4.0.0" and sys_platform != "win32"
numpy==2.0.2 ; python_version >= "3.9" and python_version < "3.11"
numpy==2.3.3 ; python_version >= "3.11" and python_full_version < "4.0.0"
orjson==3.11.3 ; python_version >= "3.9" and python_full_version < "4.0.0" and platform_python_implementation != "PyPy"
packaging==25.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
pandas==2.3.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
protobuf==6.32.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
pydantic-core==2.33.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
pydantic==2.11.7 ; python_version >= "3.9" and python_full_version < "4.0.0"
pymilvus==2.6.1 ; python_version >= "3.9" and python_full_version < "4.0.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.9" and python_full_version < "4.0.0"
python-dotenv==1.1.1 ; python_version >= "3.9" and python_full_version < "4.0.0"
pytz==2025.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
requests-toolbelt==1.0.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
requests==2.32.5 ; python_version >= "3.9" and python_full_version < "4.0.0"
setuptools==80.9.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
six==1.17.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_full_version < "4.0.0"
tenacity==9.1.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
tqdm==4.67.1 ; python_version >= "3.9" and python_full_version < "4.0.0" and sys_platform != "win32"
typing-extensions==4.15.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
typing-inspection==0.4.1 ; python_version >= "3.9" and python_full_version < "4.0.0"
tzdata==2025.2 ; python_version >= "3.9" and python_full_version < "4.0.0"
ujson==5.11.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
urllib3==2.5.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
zstandard==0.24.0 ; python_version >= "3.9" and python_full_version < "4.0.0"
